refactor: ユーザー登録処理をUseCaseにまとめた

+++ src/app/UseCase/SignUp/CompleteUserRegister/Interactor.php
@@ -4,9 +4,13 @@ namespace App\UseCase\SignUp\CompleteUserRegister;
 
 use App\Infrastructure\Dao\UserDao;
 use App\Infrastructure\Dao\UserRegisterCertificationCodeDao;
+use App\UseCase\SignUp\CompleteUserRegister\Input;
+use App\UseCase\SignUp\CompleteUserRegister\Output;
 
 final class Interactor
 {
+    private const COMPLETE_MESSAGE = '登録が完了しました。';
+    private const NOT_COMPLETE_MESSAGE = '登録が出来ませんでした。';
     private Input $input;
 
     public function __construct(Input $input)
@@ -16,13 +20,15 @@ final class Interactor
         $this->userRegisterCertificationCodeDao = new UserRegisterCertificationCodeDao();
     }
 
-    public function handler()
+    public function handler(): Output
     {
-        $result = $this->insertUserInfo();
-        $this->deleteByRegisterCertificationCode();
-        return $result;
-    }
+        if ($this->insertUserInfo()) {
+            $this->deleteByRegisterCertificationCode();
+            return new Output(true, self::COMPLETE_MESSAGE);
+        }
 
+        return new Output(false, self::NOT_COMPLETE_MESSAGE);
+    }
     private function createPasswordHash(): string
     {
         return password_hash($this->input->password(), PASSWORD_DEFAULT);
@@ -37,7 +43,7 @@ final class Interactor
         );
     }
 
-    private function deleteByRegisterCertificationCode()
+    private function deleteByRegisterCertificationCode(): void
     {
         $this->userRegisterCertificationCodeDao->deleteByRegisterCertificationCode(
             $this->input->hashCertificateRegisterCode()
